otel-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=otel-requirements.txt otel-requirements.in
#
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
calver==2022.6.26
    # via -r otel-requirements.in
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
deprecated==1.2.14
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
editables==0.5
    # via hatchling
googleapis-common-protos==1.60.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.57.0
    # via opentelemetry-exporter-otlp-proto-grpc
hatchling==1.18.0
    # via -r otel-requirements.in
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via opentelemetry-api
opentelemetry-api==1.19.0
    # via
    #   -r otel-requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-sdk
    #   otel-extensions
opentelemetry-exporter-otlp==1.19.0
    # via -r otel-requirements.in
opentelemetry-exporter-otlp-proto-common==1.19.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.19.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.19.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.40b0
    # via opentelemetry-instrumentation-requests
opentelemetry-instrumentation-requests==0.40b0
    # via -r otel-requirements.in
opentelemetry-proto==1.19.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.19.0
    # via
    #   -r otel-requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   otel-extensions
opentelemetry-semantic-conventions==0.40b0
    # via
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-sdk
opentelemetry-util-http==0.40b0
    # via opentelemetry-instrumentation-requests
otel-extensions==0.2.5
    # via -r otel-requirements.in
packaging==23.1
    # via hatchling
pathspec==0.11.2
    # via hatchling
pluggy==1.3.0
    # via hatchling
protobuf==4.24.2
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==1.10.12
    # via otel-extensions
requests==2.31.0
    # via opentelemetry-exporter-otlp-proto-http
tomli==2.0.1
    # via hatchling
trove-classifiers==2023.8.7
    # via hatchling
typing-extensions==4.7.1
    # via
    #   opentelemetry-sdk
    #   pydantic
urllib3==1.26.16
    # via
    #   -r otel-requirements.in
    #   requests
wrapt==1.15.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
zipp==3.16.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
